+++ pkg/middleware/requesttermination/req_termination.go
@@ -3,6 +3,7 @@ package requesttermination
 import (
 	"context"
 	"errors"
+	"fmt"
 
 	"github.com/cloudwego/hertz/pkg/app"
 	"github.com/go-viper/mapstructure/v2"
@@ -43,22 +44,16 @@ func (m *RequestTerminationMiddleware) ServeHTTP(ctx context.Context, c *app.Req
 }
 
 func init() {
-	_ = middleware.RegisterMiddleware("request_termination", func(params map[string]any) (app.HandlerFunc, error) {
+	_ = middleware.RegisterMiddleware("request_termination", func(params any) (app.HandlerFunc, error) {
+		if params == nil {
+			return nil, errors.New("requesttermination middleware params is empty or invalid")
+		}
+
 		opts := &Options{}
 
-		config := &mapstructure.DecoderConfig{
-			Metadata: nil,
-			Result:   opts,
-			TagName:  "mapstructure",
-		}
-
-		decoder, err := mapstructure.NewDecoder(config)
+		err := mapstructure.Decode(params, &opts)
 		if err != nil {
-			return nil, err
-		}
-
-		if err := decoder.Decode(params); err != nil {
-			return nil, err
+			return nil, fmt.Errorf("requesttermination middleware params is invalid: %w", err)
 		}
 
 		if opts.StatusCode == 0 {
@@ -66,7 +61,6 @@ func init() {
 		}
 
 		m := NewMiddleware(*opts)
-
 		return m.ServeHTTP, nil
 	})
 }