+++ pkg/middleware/setvars/setvars_test.go
@@ -13,8 +13,11 @@ import (
 func TestSetVarsMiddleware(t *testing.T) {
 	h := middleware.FindHandlerByType("setvars")
 
-	params := map[string]any{
-		variable.HTTPRoute: "/orders/{order_id}",
+	params := []any{
+		map[string]any{
+			"Key":   variable.HTTPRoute,
+			"Value": "/orders/{order_id}",
+		},
 	}
 
 	m, err := h(params)
@@ -29,13 +32,23 @@ func TestSetVarsMiddleware(t *testing.T) {
 	assert.Equal(t, "/orders/{order_id}", hzCtx.GetString(variable.HTTPRoute))
 }
 
-
 func TestSetVarsWithDirectivesMiddleware(t *testing.T) {
 	h := middleware.FindHandlerByType("setvars")
 
-	params := map[string]any{
-		variable.HTTPRoute: "/orders/{order_id}",
-		"user_id": "$http.request.header.user_id",
+	params := []any{
+		map[string]any{
+			"Key":   variable.HTTPRoute,
+			"Value": "/orders/{order_id}",
+		},
+		map[string]any{
+			"Key":   "user_id",
+			"Value": "$http.request.header.user_id",
+		},
+		map[string]any{
+			"Key":     "old",
+			"Value":   "$http.request.header.old",
+			"Default": "default_value",
+		},
 	}
 
 	m, err := h(params)
@@ -43,6 +56,7 @@ func TestSetVarsWithDirectivesMiddleware(t *testing.T) {
 
 	ctx := context.Background()
 	hzCtx := app.NewContext(0)
+	hzCtx.Set("old", "123")
 	hzCtx.Request.SetMethod("GET")
 	hzCtx.Request.URI().SetPath("/orders/123")
 	hzCtx.Request.Header.Set("user_id", "996")
@@ -50,4 +64,5 @@ func TestSetVarsWithDirectivesMiddleware(t *testing.T) {
 
 	assert.Equal(t, "/orders/{order_id}", hzCtx.GetString(variable.HTTPRoute))
 	assert.Equal(t, "996", variable.GetString("$var.user_id", hzCtx))
+	assert.Equal(t, "default_value", variable.GetString("$var.old", hzCtx))
 }